+++ NEWS.md
@@ -1,3 +1,6 @@
+## gDRutils 1.7.6 - 2025-05-13
+* add `append` arg to `.set_SE_metadata`
+
 ## gDRutils 1.7.5 - 2025-05-12
 * improve logic in `cap_assay_infinities`